Add custom y axis labeling  + fix comments
Added new class property self.ylabel, a string that is used for fig.ax.set_ylabel() to set the the y axis's label.  This is an empty string '' by default, but it can be set with the function ED.set_yName(obj self,String label) - note that as with the other functions, obj can be neglected in code using this function and the python engine will know that self is implicit.

Also improved comments and fixed incorrect labeling in the plot() fxn